Refactor IntervalCache to use HashMap and half-open intervals
Java's TreeMap requires O(log N) time for get() and put() operations. However, because IntervalCache computes the key value, it is more efficient to use a HashMap for those operations. This comes at the expense of making headMax() and refresh() O(N) operations. Because put() is called on every message received by a User, having O(1) complexity for put() outweighs the O(N) complexity for headMax() and refresh().

Previously, I think I had assumed that intervals were closed. However, for a timestamp that occurs at exactly on a key, it is ambiguous as to which interval it belongs. Each interval is now half-closed such that the start time is included and the end time is not.

Signed-off-by: rtatton <rdt17@protonmail.com>