Simplify ExperimentState.Builder
Remove the need for builder interfaces with `Setter` enum. I did not make the connection before that I could just rely on the natural ordering of the enum and the use of an `EnumMap` to ensure that the Builder factory methods are called in the correct order. This allows for the same level of flexibility as before but with ~200 fewer lines of code.

Signed-off-by: rtatton <rdt17@protonmail.com>