Refactor send condition
Evaluate the send condition for each contact, rather than across contacts. If we keep a "global" (i.e., for all a user's contacts) send threshold, then it is possible that we skip a contact due to the time constraint, update our threshold, and then fail to send a message that should have been sent to the previously skipped contact. Pushing the send condition to the contacts, however, adds the complexity of needing to keep track of when to replace each contact threshold. Moreover, it appears that a higher send coefficient will be necessary than when using it in the global context. Rather than 0.6, an optimal send coefficient might be closer to 0.8 or 0.9. Will need to evaluate to confirm.

Signed-off-by: rtatton <rdt17@protonmail.com>