feat: add LastEvent for Runtimes handler
Each User now tracks the timestamp of the last event they log, regardless of whether that call to the logger results in the event being written to the event log. When the actor system is shutting down, each User logs a LastEvent record containing this timestamp. Previously, the Runtimes Handler assumed that UserEvents were present in the event log to determine the MessagePassingRuntime. However, this becomes increasingly impractical as the size of the ContactNetwork grows. With the LastEvent being logged, it is still possible to disable all other UserEvents from being logged, but still determine the MessagePassingRuntime accurately. This also does not add overhead to an instance of RiskPropagation since the LastEvents are not logged until the User receives the PostStop signal, which is sent to the User when its stopped, i.e., when the Monitor has received an IdleTimeout message at least once from every User. With 100K Users, the event log is still less than 10MB, assuming no other UserEvents are logged.